requirements.txt
asgiref==3.5.0
Brotli==1.1.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.2
cssselect2==0.7.0
defusedxml==0.7.1
Django==4.0.3
django-cleanup==6.0.0
django-cors-headers==3.11.0
django-crispy-forms==1.14.0
django-templated-mail==1.1.1
djangorestframework==3.13.1
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
fonttools==4.44.0
gunicorn==20.1.0
html5lib==1.1
idna==3.3
itypes==1.2.0
Jinja2==3.1.1
MarkupSafe==2.1.1
oauthlib==3.2.0
Pillow==10.1.0
psycopg2-binary==2.9.3
pycparser==2.21
pydyf==0.8.0
PyJWT==2.3.0
pyphen==0.14.0
python-decouple==3.6
python3-openid==3.2.0
pytz==2022.1
requests==2.27.1
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.2.0
sqlparse==0.4.2
tinycss2==1.2.1
uritemplate==4.1.1
urllib3==1.26.9
weasyprint==60.1
webencodings==0.5.1
zopfli==0.2.3
